refactor(runner): extract concurrency, prompt-prep, response-cache, binary-discovery (#109)
Splits src/gemini-runner.ts (1,317 LOC, 5 concerns) into focused modules
without behavior change:

- src/concurrency.ts (61 LOC): Semaphore, SemaphoreTimeoutError,
  MAX_CONCURRENT, QUEUE_TIMEOUT_MS, semaphore singleton
- src/binary-discovery.ts (59 LOC): discoverGeminiBinary, GEMINI_BINARY
- src/prompt-prep.ts (238 LOC): @file extraction, expandFileRefs,
  countFileRefs, LARGE_PROMPT_THRESHOLD
- src/response-cache.ts (25 LOC): cache Map, cacheKey, clearCache,
  CACHE_TTL_MS, CACHE_MAX_ENTRIES

gemini-runner.ts shrinks to 963 LOC and re-exports the moved symbols at
their original paths so the 9 import sites (resources, index, tools/*,
setup) need no changes. Module-level side-effect order is preserved by
import order in the runner: GEMINI_BINARY init -> orphan reaper kickoff
-> warm-pool instantiation -> capability detection.

tests/gemini-runner-binary-discovery.test.ts -> tests/binary-discovery.test.ts;
the discoverGeminiBinary describe block now imports from binary-discovery.js
directly to validate the new module standalone. Other describes still
import via gemini-runner.js since they test runner-level integration.

660 tests still pass; npm run build + npx tsc --noEmit clean.
